update to Meteor 2.14. Correct typeo in New Pattern button. Eslint and Prettier tweaks. First steps in replacing Futures.
