+++ .meteor/packages
@@ -5,22 +5,22 @@
 # but you can also edit it by hand.
 
 meteor-base@1.5.1             # Packages every Meteor app needs to have
-mobile-experience@1.1.0       # Packages for a great mobile UX
-mongo@1.16.3                   # The database Meteor supports right now
+mobile-experience@1.1.1       # Packages for a great mobile UX
+mongo@1.16.8                   # The database Meteor supports right now
 blaze-html-templates@1.0.4 # Compile .html files into Meteor Blaze views
-tracker@1.2.1                 # Meteor's client-side reactive programming library
+tracker@1.3.3                 # Meteor's client-side reactive programming library
 
-standard-minifier-css@1.8.3   # CSS minifier run for production mode
+standard-minifier-css@1.9.2   # CSS minifier run for production mode
 standard-minifier-js@2.8.1    # JS minifier run for production mode
 es5-shim@4.8.0                # ECMAScript 5 compatibility for older browsers
-ecmascript@0.16.4              # Enable ECMAScript2015+ syntax in app code
+ecmascript@0.16.8              # Enable ECMAScript2015+ syntax in app code
 shell-server@0.5.0            # Server-side component of the `meteor shell` command
 
 check@1.3.2
 fourseven:scss
 tmeasday:publish-counts
-accounts-password@2.3.2
-email@2.2.3
+accounts-password@2.4.0
+email@2.2.5
 meteortesting:mocha
 dburles:factory
 johanbrook:publication-collector
