+++ .meteor/versions
@@ -1,45 +1,45 @@
-accounts-base@2.2.6
-accounts-password@2.3.2
-alanning:roles@3.4.0
+accounts-base@2.2.9
+accounts-password@2.4.0
+alanning:roles@3.6.1
 aldeed:collection2@3.5.0
 aldeed:schema-index@3.0.0
 allow-deny@1.1.1
 autoupdate@1.8.0
-babel-compiler@7.10.1
+babel-compiler@7.10.5
 babel-runtime@1.5.1
 base64@1.0.12
 binary-heap@1.0.11
-blaze@2.6.1
+blaze@2.7.1
 blaze-html-templates@1.2.1
 blaze-tools@1.1.3
-boilerplate-generator@1.7.1
+boilerplate-generator@1.7.2
 caching-compiler@1.2.2
 caching-html-compiler@1.2.1
-callback-hook@1.4.0
+callback-hook@1.5.1
 check@1.3.2
 coffeescript@2.4.1
 coffeescript-compiler@2.4.1
-dburles:factory@1.3.0
+dburles:factory@1.5.0
 ddp@1.4.1
 ddp-client@2.6.1
 ddp-common@1.4.0
-ddp-rate-limiter@1.1.1
-ddp-server@2.6.0
+ddp-rate-limiter@1.2.1
+ddp-server@2.7.0
 diff-sequence@1.1.2
-dynamic-import@0.7.2
+dynamic-import@0.7.3
 easy:search@2.2.1
 easysearch:components@2.2.2
 easysearch:core@2.2.2
-ecmascript@0.16.4
-ecmascript-runtime@0.8.0
+ecmascript@0.16.8
+ecmascript-runtime@0.8.1
 ecmascript-runtime-client@0.12.1
 ecmascript-runtime-server@0.11.0
 edgee:slingshot@0.7.1
 ejson@1.1.3
-email@2.2.3
+email@2.2.5
 es5-shim@4.8.0
-fetch@0.1.2
-fourseven:scss@4.15.0
+fetch@0.1.4
+fourseven:scss@4.16.0
 geojson-utils@1.0.11
 hot-code-push@1.0.4
 html-tools@1.1.3
@@ -49,30 +49,30 @@ id-map@1.1.1
 inter-process-messaging@0.1.1
 johanbrook:publication-collector@1.1.0
 jquery@1.11.11
-launch-screen@1.3.0
+launch-screen@2.0.0
 lmieulet:meteor-coverage@3.2.0
 localstorage@1.2.0
-logging@1.3.1
-matb33:collection-hooks@1.1.4
-meteor@1.10.3
+logging@1.3.3
+matb33:collection-hooks@1.3.0
+meteor@1.11.4
 meteor-base@1.5.1
-meteortesting:browser-tests@1.5.1
+meteortesting:browser-tests@1.5.3
 meteortesting:mocha@1.1.5
 meteortesting:mocha-core@7.2.0
-minifier-css@1.6.2
+minifier-css@1.6.4
 minifier-js@2.7.5
-minimongo@1.9.1
-mobile-experience@1.1.0
+minimongo@1.9.3
+mobile-experience@1.1.1
 mobile-status-bar@1.1.0
-modern-browsers@0.1.9
-modules@0.19.0
+modern-browsers@0.1.10
+modules@0.20.0
 modules-runtime@0.13.1
-mongo@1.16.3
+mongo@1.16.8
 mongo-decimal@0.1.3
 mongo-dev-server@1.1.0
 mongo-id@1.0.8
-npm-mongo@4.12.1
-observe-sequence@1.0.20
+npm-mongo@4.17.2
+observe-sequence@1.0.21
 ordered-dict@1.1.0
 peerlibrary:assert@0.3.0
 peerlibrary:base-component@0.17.1
@@ -83,9 +83,9 @@ peerlibrary:reactive-field@0.6.0
 promise@0.12.2
 raix:eventemitter@1.0.0
 random@1.2.1
-rate-limit@1.0.9
-react-fast-refresh@0.2.3
-react-meteor-data@2.6.0
+rate-limit@1.1.1
+react-fast-refresh@0.2.8
+react-meteor-data@2.7.2
 reactive-dict@1.3.1
 reactive-var@1.0.12
 reload@1.3.1
@@ -93,22 +93,23 @@ retry@1.1.0
 routepolicy@1.1.1
 sha@1.0.9
 shell-server@0.5.0
-socket-stream-client@0.5.0
+socket-stream-client@0.5.2
 spacebars@1.3.0
 spacebars-compiler@1.3.1
-standard-minifier-css@1.8.3
+standard-minifier-css@1.9.2
 standard-minifier-js@2.8.1
 templating@1.4.2
 templating-compiler@1.4.1
-templating-runtime@1.6.1
+templating-runtime@1.6.3
 templating-tools@1.2.2
 tmeasday:check-npm-versions@1.0.2
 tmeasday:publish-counts@0.8.0
-tracker@1.2.1
-typescript@4.6.4
+tracker@1.3.3
+typescript@4.9.5
 ui@1.0.13
-underscore@1.0.11
+underscore@1.0.13
 url@1.3.2
-webapp@1.13.2
+webapp@1.13.6
 webapp-hashing@1.1.1
 xolvio:cleaner@0.4.0
+zodern:types@1.0.11
